Fix some compiler warnings in libcineon and libdpx
These warnings arise with g++-4.4.1.  Some of seem to represent bugs;
- tuncation in libcineon's BaseTypeConverter,
- very odd usage for IdentifyFile

The patch also removes some spurious svn:executable properties.


git-svn-id: http://svn.openimageio.org/oiio/trunk@1617 31904243-a364-42ac-a3db-9305396518ba

+++ src/cineon.imageio/libcineon/Reader.cpp
@@ -126,7 +126,7 @@ bool cineon::Reader::ReadBlock(void *data, const DataSize size, Block &block)
 			if (!consistentWidth)
 				break;
 		}
-		if (this->header.PixelsPerLine(i) != width) {
+		if ((int)this->header.PixelsPerLine(i) != width) {
 			consistentWidth = false;
 			if (!consistentDepth)
 				break;
